rna_manual_reference_updater: disallow malformed URL's
There were no sanity checks for bad names from the manual,
this caused errors from spaces, dashes or '#' character which
aren't going to properly pattern match.

In the case of spaces, this broke string splitting logic, so skip these
names and warn. The names need to be corrected in the manual.